Frontend

+++ backend/frogcrew/src/main/java/edu/tcu/cs/frogcrew/availability/dto/AvailabilityDTO.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record AvailabilityDTO(
         @NotNull(message = "Crew Member id is required")
-        Integer crewMemberId,
+        Long crewMemberId,
 
         @NotNull(message = "Game id is required")
         Integer gameId,
